+++ tools/build-packs.mjs
@@ -1,7 +1,7 @@
 import {rm} from 'node:fs/promises';
 import {compilePack} from '@foundryvtt/foundryvtt-cli';
 
-const packs = ['gac-features-2014', 'gac-summons-2014'];
+const packs = ['gac-features-2014', 'gac-summons-2014', 'gac-equipment-2014'];
 
 for (const pack of packs) {
   const output = `./packs/${pack}`;